Garmin: Persist respiratory rate

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/messages/FitRespirationRate.java
@@ -21,8 +21,8 @@ public class FitRespirationRate extends RecordData {
     }
 
     @Nullable
-    public Integer getRespirationRate() {
-        return (Integer) getFieldByNumber(0);
+    public Float getRespirationRate() {
+        return (Float) getFieldByNumber(0);
     }
 
     @Nullable
